СПРИНТ 11 commit 2 (19.08.25 20:55)
