+++ app/src/main/java/com/saturnnetwork/playlistmaker/data_class/SearchResponse.kt
@@ -1,0 +1,8 @@
+package com.saturnnetwork.playlistmaker.data_class
+
+import com.saturnnetwork.playlistmaker.Track
+
+data class SearchResponse(
+    val resultCount: Int,
+    val results: ArrayList<Track>
+)
